Move definitions to debugger.entities
- this unifies ActivityType and ActivityDef, which were split for no good reason
- brings more uniformity in general
