+++ src/som/primitives/ActivitySpawn.java
@@ -38,8 +38,8 @@ import som.vmobjects.SSymbol;
 import tools.concurrency.ActorExecutionTrace;
 import tools.concurrency.Tags.ActivityCreation;
 import tools.concurrency.Tags.ExpressionBreakpoint;
+import tools.debugger.entities.ActivityType;
 import tools.debugger.entities.BreakpointType;
-import tools.debugger.entities.TraceSemantics.ActivityDef;
 import tools.debugger.nodes.AbstractBreakpointNode;
 import tools.debugger.session.Breakpoints;
 
@@ -51,7 +51,7 @@ public abstract class ActivitySpawn {
     SomForkJoinTask task;
     if (VmSettings.ACTOR_TRACING) {
       task = new TracedForkJoinTask(argArray, stopOnRoot);
-      ActorExecutionTrace.activityCreation(ActivityDef.TASK, task.getId(),
+      ActorExecutionTrace.activityCreation(ActivityType.TASK, task.getId(),
           block.getMethod().getSignature(), section);
     } else {
       task = new SomForkJoinTask(argArray, stopOnRoot);
@@ -64,7 +64,7 @@ public abstract class ActivitySpawn {
     SomThreadTask thread;
     if (VmSettings.ACTOR_TRACING) {
       thread = new TracedThreadTask(argArray, stopOnRoot);
-      ActorExecutionTrace.activityCreation(ActivityDef.THREAD, thread.getId(),
+      ActorExecutionTrace.activityCreation(ActivityType.THREAD, thread.getId(),
           block.getMethod().getSignature(), section);
     } else {
       thread = new SomThreadTask(argArray, stopOnRoot);
@@ -76,7 +76,7 @@ public abstract class ActivitySpawn {
       final SourceSection origin, final boolean stopOnRoot) {
     if (VmSettings.ACTOR_TRACING) {
       TracingProcess result = new TracingProcess(obj, stopOnRoot);
-      ActorExecutionTrace.activityCreation(ActivityDef.PROCESS,
+      ActorExecutionTrace.activityCreation(ActivityType.PROCESS,
           result.getId(), result.getProcObject().getSOMClass().getName(), origin);
       return result;
     } else {
