+++ src/som/primitives/actors/CreateActorPrim.java
@@ -18,7 +18,7 @@ import som.vmobjects.SClass;
 import tools.concurrency.ActorExecutionTrace;
 import tools.concurrency.Tags.ActivityCreation;
 import tools.concurrency.Tags.ExpressionBreakpoint;
-import tools.debugger.entities.TraceSemantics.ActivityDef;
+import tools.debugger.entities.ActivityType;
 
 
 @GenerateNodeFactory
@@ -43,7 +43,7 @@ public abstract class CreateActorPrim extends BinaryComplexOperation {
     if (VmSettings.ACTOR_TRACING) {
       assert argument instanceof SClass;
       final SClass actorClass = (SClass) argument;
-      ActorExecutionTrace.activityCreation(ActivityDef.ACTOR, actor.getId(),
+      ActorExecutionTrace.activityCreation(ActivityType.ACTOR, actor.getId(),
           actorClass.getName(), sourceSection);
     }
     return ref;
