+++ src/tools/debugger/entities/ActivityType.java
@@ -1,22 +1,42 @@
 package tools.debugger.entities;
 
+import tools.TraceData;
+
+
 public enum ActivityType {
-  PROCESS(EntityType.PROCESS, "&#10733;"),
-  ACTOR(EntityType.ACTOR,     "&#128257;"),
-  TASK(EntityType.TASK,       "&#8623;"),
-  THREAD(EntityType.THREAD,   "&#11123;");
+  PROCESS(EntityType.PROCESS, "&#10733;",  Marker.PROCESS_CREATION, Marker.PROCESS_COMPLETION),
+  ACTOR(EntityType.ACTOR,     "&#128257;", Marker.ACTOR_CREATION),
+  TASK(EntityType.TASK,       "&#8623;",   Marker.TASK_SPAWN),
+  THREAD(EntityType.THREAD,   "&#11123;",  Marker.THREAD_SPAWN);
 
   private final EntityType type;
-  private final String marker;
+  private final String icon;
 
-  ActivityType(final EntityType type, final String marker) {
+  private final byte creationMarker;
+  private final byte completionMarker;
+
+  ActivityType(final EntityType type, final String icon,
+      final byte creationMarker, final byte completionMarker) {
     this.type = type;
-    this.marker = marker;
+    this.icon = icon;
+    this.creationMarker   = creationMarker;
+    this.completionMarker = completionMarker;
+  }
+
+  ActivityType(final EntityType type, final String icon,
+      final byte creationMarker) {
+    this(type, icon, creationMarker, (byte) 0);
   }
 
   public EntityType getType() { return type; }
   public String getName() { return type.name; }
 
   public byte getId() { return type.id; }
-  public String getMarker() { return marker; }
+  public String getIcon() { return icon; }
+
+  public byte getCreationMarker()   { return creationMarker; }
+  public byte getCompletionMarker() { return completionMarker; }
+
+  public int getCreationSize()   { return 11 + TraceData.SOURCE_SECTION_SIZE; }
+  public int getCompletionSize() { return 1; }
 }
